# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SplitSmart — Creative Expense Splitter API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface Group {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  currency: string;
  category: string;
  emoji?: string;
  createdAt: string;
  updatedAt?: string;
}

export interface GroupInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minLength 1 */
  currency: string;
  category: string;
  emoji?: string;
}

export interface GroupUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  currency?: string;
  category?: string;
  emoji?: string;
}

export interface CategoryStat {
  category: string;
  total: number;
  count: number;
}

export interface GroupSummary {
  groupId: number;
  totalSpent: number;
  expenseCount: number;
  memberCount: number;
  categoryBreakdown: CategoryStat[];
  /** @nullable */
  topSpender: string | null;
  /** @nullable */
  mostExpensiveCategory?: string | null;
}

export interface MemberBalance {
  memberId: number;
  memberName: string;
  avatarColor?: string;
  paid: number;
  owes: number;
  netBalance: number;
}

export interface SettlementSuggestion {
  fromMemberId: number;
  fromMemberName: string;
  toMemberId: number;
  toMemberName: string;
  amount: number;
}

export interface Member {
  id: number;
  groupId: number;
  name: string;
  /** @nullable */
  email?: string | null;
  avatarColor: string;
  createdAt: string;
}

export interface MemberInput {
  /** @minLength 1 */
  name: string;
  email?: string;
  avatarColor?: string;
}

export interface MemberUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  avatarColor?: string;
}

export interface Expense {
  id: number;
  groupId: number;
  paidById: number;
  title: string;
  amount: number;
  category: string;
  splitType: string;
  /** @nullable */
  notes?: string | null;
  date: string;
  createdAt: string;
}

export interface ExpenseSplit {
  id: number;
  expenseId: number;
  memberId: number;
  amount: number;
  /** @nullable */
  percentage?: number | null;
}

export interface ExpenseDetail {
  id: number;
  groupId: number;
  paidById: number;
  title: string;
  amount: number;
  category: string;
  splitType: string;
  /** @nullable */
  notes?: string | null;
  date: string;
  createdAt: string;
  splits: ExpenseSplit[];
}

export interface SplitInput {
  memberId: number;
  amount?: number;
  percentage?: number;
}

export interface ExpenseInput {
  paidById: number;
  /** @minLength 1 */
  title: string;
  /** @minimum 0.01 */
  amount: number;
  category: string;
  splitType: string;
  notes?: string;
  date: string;
  splits?: SplitInput[];
}

export interface ExpenseUpdate {
  paidById?: number;
  /** @minLength 1 */
  title?: string;
  /** @minimum 0.01 */
  amount?: number;
  category?: string;
  notes?: string;
  date?: string;
}

export interface Settlement {
  id: number;
  groupId: number;
  fromMemberId: number;
  toMemberId: number;
  amount: number;
  /** @nullable */
  note?: string | null;
  settledAt: string;
}

export interface SettlementInput {
  fromMemberId: number;
  toMemberId: number;
  /** @minimum 0.01 */
  amount: number;
  note?: string;
}