# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SplitSmart — Creative Expense Splitter API
 * OpenAPI spec version: 1.0.0
 */

export interface Expense {
  id: number;
  groupId: number;
  paidById: number;
  title: string;
  amount: number;
  category: string;
  splitType: string;
  /** @nullable */
  notes?: string | null;
  date: string;
  createdAt: Date;
}