# lib/api-zod/src/generated/types/expenseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SplitSmart — Creative Expense Splitter API
 * OpenAPI spec version: 1.0.0
 */
import type { SplitInput } from './splitInput';

export interface ExpenseInput {
  paidById: number;
  /** @minLength 1 */
  title: string;
  /** @minimum 0.01 */
  amount: number;
  category: string;
  splitType: string;
  notes?: string;
  date: string;
  splits?: SplitInput[];
}