# lib/api-zod/src/generated/types/expenseSplit.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SplitSmart — Creative Expense Splitter API
 * OpenAPI spec version: 1.0.0
 */

export interface ExpenseSplit {
  id: number;
  expenseId: number;
  memberId: number;
  amount: number;
  /** @nullable */
  percentage?: number | null;
}