# lib/api-zod/src/generated/types/groupInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SplitSmart — Creative Expense Splitter API
 * OpenAPI spec version: 1.0.0
 */

export interface GroupInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  /** @minLength 1 */
  currency: string;
  category: string;
  emoji?: string;
}