# lib/api-zod/src/generated/types/groupSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SplitSmart — Creative Expense Splitter API
 * OpenAPI spec version: 1.0.0
 */
import type { CategoryStat } from './categoryStat';

export interface GroupSummary {
  groupId: number;
  totalSpent: number;
  expenseCount: number;
  memberCount: number;
  categoryBreakdown: CategoryStat[];
  /** @nullable */
  topSpender: string | null;
  /** @nullable */
  mostExpensiveCategory?: string | null;
}