# lib/api-zod/src/generated/types/groupUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SplitSmart — Creative Expense Splitter API
 * OpenAPI spec version: 1.0.0
 */

export interface GroupUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  currency?: string;
  category?: string;
  emoji?: string;
}