# lib/api-zod/src/generated/types/splitInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SplitSmart — Creative Expense Splitter API
 * OpenAPI spec version: 1.0.0
 */

export interface SplitInput {
  memberId: number;
  amount?: number;
  percentage?: number;
}